fork creation

+++ benchmark/bellmanford/frst.c
@@ -1,8 +1,8 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef float TARGET_TYPE;
-typedef unsigned long TARGET_INDEX;
+typedef int8_t TARGET_TYPE;
+typedef uint8_t TARGET_INDEX;
 
 TARGET_TYPE dist[size];
 TARGET_INDEX i, j = 0;
